refactor(promql): move the promql helper functions to internal/promql
