+++ stdlib/internal/promql/date_functions.go
@@ -37,13 +37,13 @@ func generateDateFunction(name string, dateFn func(time.Time) int) values.Functi
 }
 
 func init() {
-	flux.RegisterPackageValue("promql", "promqlDayOfMonth", generateDateFunction("promqlDayOfMonth", dayOfMonthFn))
-	flux.RegisterPackageValue("promql", "promqlDayOfWeek", generateDateFunction("promqlDayOfWeek", dayOfWeekFn))
-	flux.RegisterPackageValue("promql", "promqlDaysInMonth", generateDateFunction("promqlDaysInMonth", daysInMonthFn))
-	flux.RegisterPackageValue("promql", "promqlHour", generateDateFunction("promqlHour", hourFn))
-	flux.RegisterPackageValue("promql", "promqlMinute", generateDateFunction("promqlMinute", minuteFn))
-	flux.RegisterPackageValue("promql", "promqlMonth", generateDateFunction("promqlMonth", monthFn))
-	flux.RegisterPackageValue("promql", "promqlYear", generateDateFunction("promqlYear", yearFn))
+	flux.RegisterPackageValue("internal/promql", "promqlDayOfMonth", generateDateFunction("promqlDayOfMonth", dayOfMonthFn))
+	flux.RegisterPackageValue("internal/promql", "promqlDayOfWeek", generateDateFunction("promqlDayOfWeek", dayOfWeekFn))
+	flux.RegisterPackageValue("internal/promql", "promqlDaysInMonth", generateDateFunction("promqlDaysInMonth", daysInMonthFn))
+	flux.RegisterPackageValue("internal/promql", "promqlHour", generateDateFunction("promqlHour", hourFn))
+	flux.RegisterPackageValue("internal/promql", "promqlMinute", generateDateFunction("promqlMinute", minuteFn))
+	flux.RegisterPackageValue("internal/promql", "promqlMonth", generateDateFunction("promqlMonth", monthFn))
+	flux.RegisterPackageValue("internal/promql", "promqlYear", generateDateFunction("promqlYear", yearFn))
 }
 
 func dayOfMonthFn(t time.Time) int {
