+++ stdlib/packages.go
@@ -14,11 +14,11 @@ import (
 	_ "github.com/influxdata/flux/stdlib/influxdata/influxdb/monitor"
 	_ "github.com/influxdata/flux/stdlib/influxdata/influxdb/secrets"
 	_ "github.com/influxdata/flux/stdlib/influxdata/influxdb/v1"
+	_ "github.com/influxdata/flux/stdlib/internal/promql"
 	_ "github.com/influxdata/flux/stdlib/json"
 	_ "github.com/influxdata/flux/stdlib/kafka"
 	_ "github.com/influxdata/flux/stdlib/math"
 	_ "github.com/influxdata/flux/stdlib/pagerduty"
-	_ "github.com/influxdata/flux/stdlib/promql"
 	_ "github.com/influxdata/flux/stdlib/regexp"
 	_ "github.com/influxdata/flux/stdlib/runtime"
 	_ "github.com/influxdata/flux/stdlib/slack"
